Refactor namespace for generic compatibility (#379)
