+++ examples/bnn.py
@@ -17,7 +17,7 @@ import jax.random as random
 import numpyro
 from numpyro import handlers
 import numpyro.distributions as dist
-from numpyro.mcmc import MCMC, NUTS
+from numpyro.infer import MCMC, NUTS
 
 matplotlib.use('Agg')  # noqa: E402
