+++ examples/stochastic_volatility.py
@@ -9,7 +9,7 @@ import numpyro
 import numpyro.distributions as dist
 from numpyro.examples.datasets import SP500, load_dataset
 from numpyro.hmc_util import initialize_model
-from numpyro.mcmc import hmc
+from numpyro.infer.mcmc import hmc
 from numpyro.util import fori_collect
