+++ numpyro/compat/distributions.py
@@ -1,2 +1,1 @@
 from numpyro.distributions import *  # noqa: F401, F403
-from numpyro.distributions.constraints import *  # noqa: F401, F403
