+++ numpyro/compat/pyro.py
@@ -1,2 +1,1 @@
-import numpyro.patch  # noqa: F401
 from numpyro.primitives import module, param, plate, sample  # noqa: F401
