+++ numpyro/infer/__init__.py
@@ -1,0 +1,10 @@
+from numpyro.infer.mcmc import HMC, MCMC, NUTS
+from numpyro.infer.svi import SVI, elbo
+
+__all__ = [
+    'elbo',
+    'HMC',
+    'MCMC',
+    'NUTS',
+    'SVI',
+]
