+++ test/test_svi.py
@@ -7,9 +7,9 @@ import numpyro
 from numpyro import optim
 import numpyro.distributions as dist
 from numpyro.distributions import constraints
-from numpyro.distributions.constraints import AffineTransform, SigmoidTransform
+from numpyro.distributions.transforms import AffineTransform, SigmoidTransform
 from numpyro.handlers import substitute
-from numpyro.svi import SVI, elbo
+from numpyro.infer import SVI, elbo
 from numpyro.util import fori_loop
